proof+itest: move aperture harness to itest
This avoids including the itest package in the production build of tapd.
